backend/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
celery
redis
bcrypt
PyJWT